+++ gp-proj/src/projections/projections/traits.rs
@@ -33,9 +33,6 @@ pub struct DistortionMetrics {
 }
 
 pub trait Projection {
-    fn geo_to_face(&self, positions: Vec<Point>, polyhedron: Option<&Polyhedron>) -> Vec<ForwardBary>;
-    fn face_to_geo(&self, coords: Vec<Coord>) -> Point;
-
     fn geo_to_cartesian(
         &self,
         positions: Vec<Point>,
