+++ gp-proj/src/projections/projections/vgc.rs
@@ -14,17 +14,55 @@ use crate::{
     models::vector_3d::Vector3D,
     projections::{
         layout::traits::Layout,
-        polyhedron::{ArcLengths, Polyhedron, spherical_geometry::spherical_triangle_area},
-        projections::traits::{DistortionMetrics, ForwardBary, ForwardCartesian, Projection},
-    },
-    utils::shape::{
-        FACE_TEMPLATE_DOWN, FACE_TEMPLATE_UP, SUB_TRIANGLE_TEMPLATE, affine_transform_triangle,
-        cartesian_2d_to_barycentric, compute_spherical_barycentric,
-        get_subtriangle_vertices_in_face, map_subtriangle_vertices_to_face_2d, triangle,
-        triangle3d_to_2d,
-    },
+        polyhedron::{ArcLengths, Polyhedron},
+        projections::traits::{DistortionMetrics, ForwardCartesian, Projection},
+    }, utils::shape::triangle,
 };
 use geo::{Coord, Point};
+
+// SUB_TRIANGLE_TEMPLATE
+// Each icosahedron face is divided into 6 sub-triangles by connecting the face center
+// to the midpoints of each edge. All 6 sub-triangles are congruent right triangles.
+// Arc lengths measured on the unit sphere for one sub-triangle:
+//   ab (corner to mid) = 0.553574 rad
+//   bc (corner to center) = 0.652358 rad  
+//   ac (mid to center) = 0.364864 rad
+// Template is built with B (corner) at origin, A (mid) on negative x-axis,
+// C (center) placed using law of cosines at B:
+//   cos(angle_B) = (ab² + bc² - ac²) / (2·ab·bc)
+//   C = (bc·cos(angle_B), bc·sin(angle_B))
+// Raw planar area = 0.5 * |ab * C.y| ≈ 0.100929
+// Spherical sub-triangle area = 4π / 120 ≈ 0.104720 (unit sphere, 20 faces × 6 sub-triangles)
+// Scale factor = sqrt(0.104720 / 0.100929) ≈ 1.018606
+// All coordinates multiplied by 1.018606 to match spherical sub-triangle area.
+const SCALE_SUB: f64 = 1.018606;
+const SUB_TRIANGLE_TEMPLATE: [(f64, f64); 3] = [
+    (0.0,                           0.0),        // B = corner (origin)
+    (-0.553574 * SCALE_SUB,         0.0),        // A = mid
+    ( 0.540930 * SCALE_SUB,         0.364645 * SCALE_SUB), // C = center
+];
+// FACE_TEMPLATE_UP and FACE_TEMPLATE_DOWN
+// Edge lengths come from the regular icosahedron on a unit sphere:
+//   - edge 0-1: π/3 ≈ 1.107149 rad (exact)
+//   - edge 1-2 and 2-0: ≈ 1.107149 rad (all equal, regular icosahedron)
+// Triangle is built with f1 at origin, f0 on negative x-axis, f2 using law of cosines at f1.
+// Raw planar area = 0.5 * |(-1.107149 * 0.958819)| ≈ 0.530938
+// Spherical face area = 4π / 20 ≈ 0.628318 (unit sphere, 20 equal faces)
+// Scale factor = sqrt(0.628318 / 0.530938) ≈ 1.088072
+// All coordinates multiplied by 1.088072 to make planar area equal spherical face area,
+// ensuring the equal-area property is preserved when mapping to the face plane.
+const SCALE_FACE: f64 = 1.0880715;
+const FACE_TEMPLATE_UP: [(f64, f64); 3] = [
+    (0.0,                          0.0),
+    (-1.107149 * SCALE_FACE,       0.0),
+    (-0.553574 * SCALE_FACE,       0.958819 * SCALE_FACE),
+];
+const FACE_TEMPLATE_DOWN: [(f64, f64); 3] = [
+    (0.0,                          0.0),
+    (-1.107149 * SCALE_FACE,       0.0),
+    (-0.553574 * SCALE_FACE,      -0.958819 * SCALE_FACE),
+];
+
 
 /// Implementation for Vertex Great Circle projection (or van Leeuwen Great Circle projection).
 /// vgc - Vertex-oriented Great Circle projection.
@@ -33,246 +71,11 @@ use geo::{Coord, Point};
 pub struct Vgc;
 
 impl Projection for Vgc {
-    fn geo_to_face(
-        &self,
-        positions: Vec<Point>,
-        polyhedron: Option<&Polyhedron>,
-    ) -> Vec<ForwardBary> {
-        let mut out: Vec<ForwardBary> = vec![];
-        let polyhedron = polyhedron.unwrap();
-
-        // Need the coeficcients to convert from geodetic to authalic
-        let coef_fourier_geod_to_auth =
-            Self::fourier_coefficients(KarneyCoefficients::GEODETIC_TO_AUTHALIC);
-
-        for position in positions {
-            let lon = position.x().to_radians();
-            let lat = Self::lat_geodetic_to_authalic(
-                position.y().to_radians(),
-                &coef_fourier_geod_to_auth,
-            );
-            // Calculate 3d unit vectors for point P
-            let point_p = Vector3D::from_array(Self::to_3d(lat, lon));
-
-            // starting from here, you need:
-            // - the 3d point that you want to project
-            // Polyhedron faces
-            let faces_length = polyhedron.num_faces();
-            for index in 0..faces_length {
-                let face = usize::from(index);
-
-                if polyhedron.is_point_in_face(point_p, index) {
-                    // the icosahedron triangle gets divided into six rectangle triangles,
-                    // and we find the one where the point is
-                    let sub_triangle_3d = triangle(polyhedron, point_p, face).unwrap();
-
-                    let face_vertices_3d = polyhedron.face_vertices(face).unwrap();
-                    // calculating the arc lenghts from one of the vertices of the sub-triangle to point P
-                    let ArcLengths {
-                        ab, bp, ap, bc, ac, ..
-                    } = polyhedron.arc_lengths(sub_triangle_3d.0, point_p);
-
-                    // Parameterization values of the slice and dice projection.
-                    let [xy, uv] = slice_and_dice(ac, ab, bc, ap, bp);
-
-                    let face_edge_lengths = polyhedron.face_arc_lengths(face).unwrap();
-
-                    let is_upward = if face % 2 == 0 { true } else { false };
-
-                    // need to scale each face individually based on that specific face's spherical area
-                    let face_spherical_area = spherical_triangle_area([
-                        face_vertices_3d[0],
-                        face_vertices_3d[1],
-                        face_vertices_3d[2],
-                    ])
-                    .unwrap();
-
-                    let face_2d_vertices = triangle3d_to_2d(
-                        face_edge_lengths[0],
-                        face_edge_lengths[1],
-                        face_edge_lengths[2],
-                        is_upward,
-                        face_spherical_area,
-                    );
-                    // // Calculate barycentric coordinates within sub-triangle for P for the subtriangle ABC
-                    // // Deduction using interpolation for the slice and dice, point D and then point P
-                    // // Check the image here: https://raw.githubusercontent.com/GeoPlegma/GeoPlegma/refs/heads/master/gp-proj/src/assets/sub-triangles.png
-                    // // P = B + (D - B) * xy => P = B * (1 - xy) + D * xy
-                    // // So (1 - xy) and xy are the barycentric coordinates (the weights) of B and D respectively
-                    // // Being that D = C + (A - C) * uv => D = A * uv + C * (1 - uv) then if you replace D in the previous equation
-                    // // P = B * (1 - xy) + (A * uv + C * (1 - uv)) * xy => A × (xy × uv) + B × (1 - xy) + C × (xy × (1 - uv))
-                    // // Therefore, the barycentric coordinates are:
-                    // // - Weight for A (v_mid)**: `xy × uv`
-                    // // - Weight for B (corner)**: `1 - xy`
-                    // // - Weight for C (center)**: `xy × (1 - uv)`
-
-                    // // let subtriangle_bary_u = xy * uv; // weight for v_mid (A)
-                    // // let subtriangle_bary_v = 1.0 - xy; // weight for corner (B)
-                    // // let subtriangle_bary_w = xy * (1.0 - uv); // weight for center (C)
-
-                    // // Get barycentric coordinates of sub-triangle vertices with respect to face
-                    // let sub_a_bary = compute_spherical_barycentric(
-                    //     sub_triangle_3d.0[0], // v_mid
-                    //     face_vertices_3d[0],
-                    //     face_vertices_3d[1],
-                    //     face_vertices_3d[2],
-                    // );
-                    // // Result: A = F0 × a0 + F1 × a1 + F2 × a2
-
-                    // let sub_b_bary = compute_spherical_barycentric(
-                    //     sub_triangle_3d.0[1], // corner
-                    //     face_vertices_3d[0],
-                    //     face_vertices_3d[1],
-                    //     face_vertices_3d[2],
-                    // );
-                    // // Result: B = F0 × b0 + F1 × b1 + F2 × b2
-
-                    // let sub_c_bary = compute_spherical_barycentric(
-                    //     sub_triangle_3d.0[2], // center
-                    //     face_vertices_3d[0],
-                    //     face_vertices_3d[1],
-                    //     face_vertices_3d[2],
-                    // );
-                    // // Result: C = F0 × c0 + F1 × c1 + F2 × c2
-
-                    // // Compose: barycentric of P in face = weighted sum of sub-triangle vertices' barycentrics
-                    // // Because barycentric coordinates are linear. When you have:
-                    // // - P as a barycentric combination of A, B, C
-                    // // - And A, B, C are themselves barycentric combinations of F0, F1, F2
-                    // // You can "distribute" and get P directly as a barycentric combination of F0, F1, F2.
-                    // // So then if we substitute the face-barycentric expressions for A, B, C back into the equation for P:
-                    // // P = A × subtriangle_bary_u + B × subtriangle_bary_v + C × subtriangle_bary_w
-                    // // Becomes (being that F(0,1,2) are the corners of the face):
-                    // // P = (F0×a0 + F1×a1 + F2×a2) × subtriangle_bary_u + (F0×b0 + F1×b1 + F2×b2) × subtriangle_bary_v + (F0×c0 + F1×c1 + F2×c2) × subtriangle_bary_w
-                    // // Rearranging by grouping F0, F1, F2:
-                    // // P = F0 × (a0×subtriangle_bary_u  + b0×subtriangle_bary_v  + c0×subtriangle_bary_w ) + F1 × (a1×subtriangle_bary_u  + b1×subtriangle_bary_v  + c1×subtriangle_bary_w ) + F2 × (a2×subtriangle_bary_u  + b2×subtriangle_bary_v  + c2×subtriangle_bary_w )
-                    // // let p_bary_u = subtriangle_a_bary_face.0 * subtriangle_bary_u
-                    // //     + subtriangle_b_bary_face.0 * subtriangle_bary_v
-                    // //     + subtriangle_c_bary_face.0 * subtriangle_bary_w;
-
-                    // // let p_bary_v = subtriangle_a_bary_face.1 * subtriangle_bary_u
-                    // //     + subtriangle_b_bary_face.1 * subtriangle_bary_v
-                    // //     + subtriangle_c_bary_face.1 * subtriangle_bary_w;
-
-                    // // let p_bary_w = subtriangle_a_bary_face.2 * subtriangle_bary_u
-                    // //     + subtriangle_b_bary_face.2 * subtriangle_bary_v
-                    // //     + subtriangle_c_bary_face.2 * subtriangle_bary_w;
-                    // // Convert to 2D positions
-                    // let sub_a_2d = (
-                    //     face_2d_vertices[0].0 * sub_a_bary.0
-                    //         + face_2d_vertices[1].0 * sub_a_bary.1
-                    //         + face_2d_vertices[2].0 * sub_a_bary.2,
-                    //     face_2d_vertices[0].1 * sub_a_bary.0
-                    //         + face_2d_vertices[1].1 * sub_a_bary.1
-                    //         + face_2d_vertices[2].1 * sub_a_bary.2,
-                    // );
-                    // let sub_b_2d = (
-                    //     face_2d_vertices[0].0 * sub_b_bary.0
-                    //         + face_2d_vertices[1].0 * sub_b_bary.1
-                    //         + face_2d_vertices[2].0 * sub_b_bary.2,
-                    //     face_2d_vertices[0].1 * sub_b_bary.0
-                    //         + face_2d_vertices[1].1 * sub_b_bary.1
-                    //         + face_2d_vertices[2].1 * sub_b_bary.2,
-                    // );
-                    // let sub_c_2d = (
-                    //     face_2d_vertices[0].0 * sub_c_bary.0
-                    //         + face_2d_vertices[1].0 * sub_c_bary.1
-                    //         + face_2d_vertices[2].0 * sub_c_bary.2,
-                    //     face_2d_vertices[0].1 * sub_c_bary.0
-                    //         + face_2d_vertices[1].1 * sub_c_bary.1
-                    //         + face_2d_vertices[2].1 * sub_c_bary.2,
-                    // );
-
-                    // let subtriangle_2d = [(1.0, 0.0), (0.0, 0.0), (1.0, bc * (PI / 6.0).sin())];
-                    let sub_area = spherical_triangle_area([
-                        sub_triangle_3d.0[0],
-                        sub_triangle_3d.0[1],
-                        sub_triangle_3d.0[2],
-                    ])
-                    .unwrap();
-
-                    let subtriangle_2d = triangle3d_to_2d(ab, bc, ac, true, sub_area);
-                    // Interpolate in 2D using slice-and-dice parameters
-                    // D = C + (A - C) * uv
-                    // let pd_x = sub_c_2d.0 + (sub_a_2d.0 - sub_c_2d.0) * uv;
-                    // let pd_y = sub_c_2d.1 + (sub_a_2d.1 - sub_c_2d.1) * uv;
-                    let pd_x =
-                        subtriangle_2d[2].0 + (subtriangle_2d[1].0 - subtriangle_2d[2].0) * uv;
-                    let pd_y =
-                        subtriangle_2d[2].1 + (subtriangle_2d[1].1 - subtriangle_2d[2].1) * uv;
-
-                    // P = B + (D - B) * xy
-                    let p_x_local = subtriangle_2d[0].0 + (pd_x - subtriangle_2d[0].0) * xy;
-                    let p_y_local = subtriangle_2d[0].1 + (pd_y - subtriangle_2d[0].1) * xy;
-
-                    // STEP 4: Find where sub-triangle vertices A, B, C are in face 2D
-                    // We need to know which sub-triangle this is to map vertices correctly
-                    let (a_face_2d, b_face_2d, c_face_2d) = map_subtriangle_vertices_to_face_2d(
-                        &sub_triangle_3d.0,
-                        &face_vertices_3d,
-                        &face_2d_vertices,
-                    );
-
-                    // STEP 5: Transform point from sub-triangle local to face coordinates
-                    // Use affine transformation based on triangle correspondence
-                    let p_face_2d = affine_transform_triangle(
-                        (p_x_local, p_y_local),
-                        subtriangle_2d,                    // source triangle [B, A, C]
-                        [b_face_2d, a_face_2d, c_face_2d], // destination triangle
-                    );
-
-                    // println!("Point 2D: ({:.6}, {:.6})", p_x, p_y);
-                    // Convert 2D Cartesian to barycentric w.r.t. face
-                    let (bary_u, bary_v, bary_w) =
-                        cartesian_2d_to_barycentric(p_face_2d, face_2d_vertices);
-
-                    // Validate
-                    const EPSILON: f64 = -1e-6;
-                    if bary_u < EPSILON || bary_v < EPSILON || bary_w < EPSILON {
-                        // println!(
-                        //     "WARNING: Negative barycentric! Point ({}, {}), Face {}, Bary: ({}, {}, {})",
-                        //     position.x(),
-                        //     position.y(),
-                        //     face,
-                        //     bary_u,
-                        //     bary_v,
-                        //     bary_w
-                        // );
-                        continue;
-                    }
-
-                    let sum: f64 = bary_u + bary_v + bary_w;
-                    if (sum - 1.0).abs() > 1e-6 {
-                        println!("WARNING: Barycentric sum {} != 1.0", sum);
-                        continue;
-                    }
-
-                    out.push(ForwardBary {
-                        coords: Vector3D {
-                            x: bary_u,
-                            y: bary_v,
-                            z: bary_w,
-                        },
-                        face: index,
-                    });
-
-                    // in case the point is on the edge of two faces, we return the first face.
-                    break;
-                }
-            }
-        }
-
-        out
-    }
-    fn face_to_geo(&self, positions: Vec<Coord>) -> Point {
-        todo!()
-    }
-
     fn geo_to_cartesian(
         &self,
         positions: Vec<Point>,
         polyhedron: Option<&Polyhedron>,
-        layout: Option<&dyn Layout>,
+        _layout: Option<&dyn Layout>,
     ) -> Vec<ForwardCartesian> {
         let mut out: Vec<ForwardCartesian> = vec![];
         let polyhedron = polyhedron.unwrap();
@@ -304,8 +107,6 @@ impl Projection for Vgc {
                         face,
                     )
                     .unwrap();
-                    let face_vertices_3d = polyhedron.face_vertices(face).unwrap();
-
                     // calculating the arc lenghts from one of the vertices of the sub-triangle to point P
                     let ArcLengths {
                         ab, bp, ap, bc, ac, ..
@@ -313,75 +114,6 @@ impl Projection for Vgc {
 
                     // Parameterization values of the slice and dice projection.
                     let [xy, uv] = slice_and_dice(ac, ab, bc, ap, bp);
-                    /*
-                    // // Get barycentric coordinates of sub-triangle vertices with respect to face
-                    // let subtriangle_a_bary_face = compute_spherical_barycentric(
-                    //     sub_triangle_3d.0[0], // v_mid
-                    //     face_vertices_3d[0],
-                    //     face_vertices_3d[1],
-                    //     face_vertices_3d[2],
-                    // );
-                    // // Result: A = F0 × a0 + F1 × a1 + F2 × a2
-
-                    // let subtriangle_b_bary_face = compute_spherical_barycentric(
-                    //     sub_triangle_3d.0[1], // corner
-                    //     face_vertices_3d[0],
-                    //     face_vertices_3d[1],
-                    //     face_vertices_3d[2],
-                    // );
-                    // // Result: B = F0 × b0 + F1 × b1 + F2 × b2
-
-                    // let subtriangle_c_bary_face = compute_spherical_barycentric(
-                    //     sub_triangle_3d.0[2], // center
-                    //     face_vertices_3d[0],
-                    //     face_vertices_3d[1],
-                    //     face_vertices_3d[2],
-                    // );
-
-                    // let face_edge_lengths = polyhedron.face_arc_lengths(face).unwrap(); */
-
-                    // let subtriangle_2d = triangle3d_to_2d(ab, bc, ac, true, sub_area);
-                    /*   // // need to scale each face individually based on that specific face's spherical area
-                          // let face_spherical_area = spherical_triangle_area([
-                          //     face_vertices_3d[0],
-                          //     face_vertices_3d[1],
-                          //     face_vertices_3d[2],
-                          // ])
-                          // .unwrap();
-
-                          // // println!("Face {}: edges={:?}, area={:.6}", face, face_edge_lengths, face_spherical_area);
-                          // let face_2d_vertices = triangle3d_to_2d(
-                          //     face_edge_lengths[0],
-                          //     face_edge_lengths[1],
-                          //     face_edge_lengths[2],
-                          //     is_upward,
-                          //     face_spherical_area,
-                          // );
-                    // println!("Face {} edge lengths: {:?}", face, face_2d_vertices);
-                          // // println!("Face 2D vertices: {:?}", face_2d_vertices);
-                          // let subtriangle_a_x = face_2d_vertices[0].0 * subtriangle_a_bary_face.0
-                          //     + face_2d_vertices[1].0 * subtriangle_a_bary_face.1
-                          //     + face_2d_vertices[2].0 * subtriangle_a_bary_face.2;
-
-                          // let subtriangle_a_y = face_2d_vertices[0].1 * subtriangle_a_bary_face.0
-                          //     + face_2d_vertices[1].1 * subtriangle_a_bary_face.1
-                          //     + face_2d_vertices[2].1 * subtriangle_a_bary_face.2;
-
-                          // let subtriangle_b_x = face_2d_vertices[0].0 * subtriangle_b_bary_face.0
-                          //     + face_2d_vertices[1].0 * subtriangle_b_bary_face.1
-                          //     + face_2d_vertices[2].0 * subtriangle_b_bary_face.2;
-
-                          // let subtriangle_b_y = face_2d_vertices[0].1 * subtriangle_b_bary_face.0
-                          //     + face_2d_vertices[1].1 * subtriangle_b_bary_face.1
-                          //     + face_2d_vertices[2].1 * subtriangle_b_bary_face.2;
-
-                          // let subtriangle_c_x = face_2d_vertices[0].0 * subtriangle_c_bary_face.0
-                          //     + face_2d_vertices[1].0 * subtriangle_c_bary_face.1
-                          //     + face_2d_vertices[2].0 * subtriangle_c_bary_face.2;
-
-                          // let subtriangle_c_y = face_2d_vertices[0].1 * subtriangle_c_bary_face.0
-                          //     + face_2d_vertices[1].1 * subtriangle_c_bary_face.1
-                          //     + face_2d_vertices[2].1 * subtriangle_c_bary_face.2;*/
 
                     // ==== Interpolation ====
                     // Between A and C it gives point D
@@ -396,7 +128,7 @@ impl Projection for Vgc {
                         SUB_TRIANGLE_TEMPLATE[1].1 + (pd_y - SUB_TRIANGLE_TEMPLATE[1].1) * xy;
                     // ======================
 
-                    let is_upward = if face % 2 == 0 { true } else { false };
+                    let is_upward = face % 2 == 0;
                     let face_template = if is_upward {
                         FACE_TEMPLATE_UP
                     } else {
@@ -415,7 +147,6 @@ impl Projection for Vgc {
                         sub_vertices_in_face,
                     );
 
-                    // STEP 5: Convert to meters
                     let r = 6371007.181;
                     out.push(ForwardCartesian {
                         coords: Coord {
@@ -433,7 +164,7 @@ impl Projection for Vgc {
         out
     }
 
-    fn cartesian_to_geo(&self, coords: Vec<Coord>) -> Point {
+    fn cartesian_to_geo(&self, _coords: Vec<Coord>) -> Point {
         todo!()
     }
 
@@ -544,6 +275,64 @@ fn slice_and_dice(ac: f64, ab: f64, bc: f64, ap: f64, bp: f64) -> [f64; 2] {
     [xy, uv]
 }
 
+/// Get the position of sub-triangle vertices in face 2D coordinates
+fn get_subtriangle_vertices_in_face(
+    sub_triangle_id: u8,
+    face_template: [(f64, f64); 3],
+) -> [(f64, f64); 3] {
+    // Face vertices
+    let [f0, f1, f2] = face_template;
+
+    // Compute face center
+    let center = ((f0.0 + f1.0 + f2.0) / 3.0, (f0.1 + f1.1 + f2.1) / 3.0);
+
+    // Compute midpoints
+    let mid_01 = ((f0.0 + f1.0) / 2.0, (f0.1 + f1.1) / 2.0);
+    let mid_12 = ((f1.0 + f2.0) / 2.0, (f1.1 + f2.1) / 2.0);
+    let mid_20 = ((f2.0 + f0.0) / 2.0, (f2.1 + f0.1) / 2.0);
+
+    // Map sub-triangle ID to its vertices [v_mid, corner, center]
+    match sub_triangle_id {
+        0 => [mid_01, f0, center], // Between f1-f2
+        1 => [mid_01, f1, center], // Between f1-f2
+        2 => [mid_12, f1, center], // Between f2-f0
+        3 => [mid_12, f2, center], // Between f2-f0
+        4 => [mid_20, f2, center], // Between f0-f1
+        5 => [mid_20, f0, center], // Between f0-f1
+        _ => panic!("Invalid sub-triangle ID"),
+    }
+}
+
+/// Affine transformation from one triangle to another
+fn affine_transform_triangle(
+    point: (f64, f64),
+    source_tri: [(f64, f64); 3],
+    dest_tri: [(f64, f64); 3],
+) -> (f64, f64) {
+    // Source vectors relative to source_tri[0]
+    let (ax, ay) = (source_tri[1].0 - source_tri[0].0, source_tri[1].1 - source_tri[0].1);
+    let (bx, by) = (source_tri[2].0 - source_tri[0].0, source_tri[2].1 - source_tri[0].1);
+
+    // Destination vectors relative to dest_tri[0]
+    let (cx, cy) = (dest_tri[1].0 - dest_tri[0].0, dest_tri[1].1 - dest_tri[0].1);
+    let (dx, dy) = (dest_tri[2].0 - dest_tri[0].0, dest_tri[2].1 - dest_tri[0].1);
+
+    // Point relative to source_tri[0]
+    let (px, py) = (point.0 - source_tri[0].0, point.1 - source_tri[0].1);
+
+    // Solve: [ax bx] [s]   [px]
+    //        [ay by] [t] = [py]
+    let det = ax * by - bx * ay;
+    let s = (px * by - bx * py) / det;
+    let t = (ax * py - px * ay) / det;
+
+    // Apply same s,t to destination
+    let x = dest_tri[0].0 + s * cx + t * dx;
+    let y = dest_tri[0].1 + s * cy + t * dy;
+
+    (x, y)
+}
+
 // @TODO - new tests need to be added.
 #[cfg(test)]
 mod tests {
@@ -576,7 +365,7 @@ mod tests {
         let projection = Vgc;
         let icosahedron = new();
         let result =
-            projection.geo_to_face(vec![p1, p2, p3, p4, p5, p6, p7, p8, p9], Some(&icosahedron));
+            projection.geo_to_cartesian(vec![p1, p2, p3, p4, p5, p6, p7, p8, p9], Some(&icosahedron), None);
 
         assert_eq!(result[0].face, 8);
         assert_eq!(result[1].face, 6);
@@ -598,7 +387,7 @@ mod tests {
         let porto = Point::new(-8.61099, 41.14961); // ~300km north of Lisbon
         let madrid = Point::new(-3.70379, 40.41678); // ~500km east of Lisbon
 
-        let results = projection.geo_to_face(vec![lisbon, porto, madrid], Some(&icosahedron));
+        let results = projection.geo_to_cartesian(vec![lisbon, porto, madrid], Some(&icosahedron), None);
 
         // Check they're on reasonable faces
         println!("Lisbon face: {}", results[0].face);
@@ -627,7 +416,7 @@ mod tests {
             Point::new(288.0, 89.0),
         ];
 
-        let results = projection.geo_to_face(points, Some(&icosahedron));
+        let results = projection.geo_to_cartesian(points, Some(&icosahedron), None);
 
         // All should be near pole (check they're on the 5 faces around the north pole)
         for (i, result) in results.iter().enumerate() {
@@ -651,7 +440,7 @@ mod tests {
         // Points evenly distributed around equator
         let points: Vec<Point> = (0..10).map(|i| Point::new(i as f64 * 36.0, 0.0)).collect();
 
-        let results = projection.geo_to_face(points, Some(&icosahedron));
+        let results = projection.geo_to_cartesian(points, Some(&icosahedron), None);
 
         // Should hit multiple different faces
         let unique_faces: std::collections::HashSet<_> = results.iter().map(|r| r.face).collect();
