+++ gp-proj/src/utils/shape.rs
@@ -11,7 +11,7 @@ use crate::{
     Vector3D,
     projections::polyhedron::{
         Polyhedron,
-        spherical_geometry::{self, spherical_triangle_area},
+        spherical_geometry::{self},
     },
 };
 
@@ -129,262 +129,4 @@ pub fn triangle(
     } else {
         Some(([mid, v1, c], (i1 * 2 + 1) as u8))
     }
-}
-
-// SUB_TRIANGLE_TEMPLATE
-// Each icosahedron face is divided into 6 sub-triangles by connecting the face center
-// to the midpoints of each edge. All 6 sub-triangles are congruent right triangles.
-// Arc lengths measured on the unit sphere for one sub-triangle:
-//   ab (corner to mid) = 0.553574 rad
-//   bc (corner to center) = 0.652358 rad  
-//   ac (mid to center) = 0.364864 rad
-// Template is built with B (corner) at origin, A (mid) on negative x-axis,
-// C (center) placed using law of cosines at B:
-//   cos(angle_B) = (ab² + bc² - ac²) / (2·ab·bc)
-//   C = (bc·cos(angle_B), bc·sin(angle_B))
-// Raw planar area = 0.5 * |ab * C.y| ≈ 0.100929
-// Spherical sub-triangle area = 4π / 120 ≈ 0.104720 (unit sphere, 20 faces × 6 sub-triangles)
-// Scale factor = sqrt(0.104720 / 0.100929) ≈ 1.018606
-// All coordinates multiplied by 1.018606 to match spherical sub-triangle area.
-const SCALE_SUB: f64 = 1.018606;
-pub const SUB_TRIANGLE_TEMPLATE: [(f64, f64); 3] = [
-    (0.0,                           0.0),        // B = corner (origin)
-    (-0.553574 * SCALE_SUB,         0.0),        // A = mid
-    ( 0.540930 * SCALE_SUB,         0.364645 * SCALE_SUB), // C = center
-];
-// FACE_TEMPLATE_UP and FACE_TEMPLATE_DOWN
-// Edge lengths come from the regular icosahedron on a unit sphere:
-//   - edge 0-1: π/3 ≈ 1.107149 rad (exact)
-//   - edge 1-2 and 2-0: ≈ 1.107149 rad (all equal, regular icosahedron)
-// Triangle is built with f1 at origin, f0 on negative x-axis, f2 using law of cosines at f1.
-// Raw planar area = 0.5 * |(-1.107149 * 0.958819)| ≈ 0.530938
-// Spherical face area = 4π / 20 ≈ 0.628318 (unit sphere, 20 equal faces)
-// Scale factor = sqrt(0.628318 / 0.530938) ≈ 1.088072
-// All coordinates multiplied by 1.088072 to make planar area equal spherical face area,
-// ensuring the equal-area property is preserved when mapping to the face plane.
-const SCALE_FACE: f64 = 1.0880715;
-pub const FACE_TEMPLATE_UP: [(f64, f64); 3] = [
-    (0.0,                          0.0),
-    (-1.107149 * SCALE_FACE,       0.0),
-    (-0.553574 * SCALE_FACE,       0.958819 * SCALE_FACE),
-];
-pub const FACE_TEMPLATE_DOWN: [(f64, f64); 3] = [
-    (0.0,                          0.0),
-    (-1.107149 * SCALE_FACE,       0.0),
-    (-0.553574 * SCALE_FACE,      -0.958819 * SCALE_FACE),
-];
-
-/// Get the position of sub-triangle vertices in face 2D coordinates
-pub fn get_subtriangle_vertices_in_face(
-    sub_triangle_id: u8,
-    face_template: [(f64, f64); 3],
-) -> [(f64, f64); 3] {
-    // Face vertices
-    let [f0, f1, f2] = face_template;
-
-    // Compute face center
-    let center = ((f0.0 + f1.0 + f2.0) / 3.0, (f0.1 + f1.1 + f2.1) / 3.0);
-
-    // Compute midpoints
-    let mid_01 = ((f0.0 + f1.0) / 2.0, (f0.1 + f1.1) / 2.0);
-    let mid_12 = ((f1.0 + f2.0) / 2.0, (f1.1 + f2.1) / 2.0);
-    let mid_20 = ((f2.0 + f0.0) / 2.0, (f2.1 + f0.1) / 2.0);
-
-    // Map sub-triangle ID to its vertices [v_mid, corner, center]
-    match sub_triangle_id {
-        0 => [mid_01, f0, center], // Between f1-f2
-        1 => [mid_01, f1, center], // Between f1-f2
-        2 => [mid_12, f1, center], // Between f2-f0
-        3 => [mid_12, f2, center], // Between f2-f0
-        4 => [mid_20, f2, center], // Between f0-f1
-        5 => [mid_20, f0, center], // Between f0-f1
-        _ => panic!("Invalid sub-triangle ID"),
-    }
-}
-// Map spherical triangle into a planar triangle.
-pub fn triangle3d_to_2d(
-    ab: f64,
-    bc: f64,
-    ac: f64,
-    is_upward: bool,
-    spherical_area: f64,
-) -> [(f64, f64); 3] {
-    // let a01 = ab; // edge 0-1
-    // let a12 = bc; // edge 1-2
-    // let a20 = ac; // edge 2-0
-    // // Build triangle with v1 at origin
-    // // v1 at origin
-    // let v1 = (0.0, 0.0);
-
-    // // v0 on negative x-axis at distance a01
-    // let v0 = (-a01, 0.0);
-
-    // // v2 positioned using law of cosines
-    // // We know: a01 (v0 to v1), a12 (v1 to v2), a20 (v2 to v0)
-    // // Find angle at v1
-    // let cos_angle_v1 = (a01.powi(2) + a12.powi(2) - a20.powi(2)) / (2.0 * a01 * a12);
-    // let angle_v1 = cos_angle_v1.clamp(-1.0, 1.0).acos();
-
-    // let y_sign = if is_upward { 1.0 } else { -1.0 };
-
-    // // v2 at distance a12 from v1, at angle from negative x-axis
-    // let v2_x = -a12 * angle_v1.cos();
-    // let v2_y = y_sign * a12 * angle_v1.sin();
-    // let v2 = (v2_x, v2_y);
-
-    // [v1, v0, v2]
-    let a01 = ab;
-    let a12 = bc;
-    let a20 = ac;
-
-    // Build triangle with edge lengths (temporarily)
-    let v1 = (0.0, 0.0);
-    let v0 = (-a01, 0.0);
-
-    let cos_angle_v1 = (a01.powi(2) + a12.powi(2) - a20.powi(2)) / (2.0 * a01 * a12);
-    let angle_v1 = cos_angle_v1.clamp(-1.0, 1.0).acos();
-
-    let y_sign = if is_upward { 1.0 } else { -1.0 };
-
-    let v2_x = -a12 * angle_v1.cos();
-    let v2_y = y_sign * a12 * angle_v1.sin();
-
-    // Calculate planar area of this triangle
-    let planar_area = 0.5 * (v0.0 * v2_y - v2_x * v0.1).abs();
-
-    // // Calculate spherical area using L'Huilier's theorem
-    // let s = (a01 + a12 + a20) / 2.0;
-    // let tan_e_over_4 = (
-    //     (s / 2.0).tan() *
-    //     ((s - a01) / 2.0).tan() *
-    //     ((s - a12) / 2.0).tan() *
-    //     ((s - a20) / 2.0).tan()
-    // ).sqrt();
-    // let spherical_area = 4.0 * tan_e_over_4.atan();
-
-    // Scale factor to make planar area equal spherical area
-    let scale = (spherical_area / planar_area).sqrt();
-
-    // Scale all vertices
-    let v0_scaled = (v0.0 * scale, v0.1 * scale);
-    let v1_scaled = (v1.0 * scale, v1.1 * scale);
-    let v2_scaled = (v2_x * scale, v2_y * scale);
-
-    [v1_scaled, v0_scaled, v2_scaled]
-}
-
-/// Map sub-triangle vertices (in 3D) to face 2D coordinate system
-pub fn map_subtriangle_vertices_to_face_2d(
-    sub_vertices_3d: &[Vector3D; 3], // [A, B, C]
-    face_vertices_3d: &[Vector3D],   // Face vertices
-    face_2d: &[(f64, f64); 3],       // Face 2D template
-) -> ((f64, f64), (f64, f64), (f64, f64)) {
-    // For each sub-triangle vertex, compute its position in face 2D
-    let mut result = [(0.0, 0.0); 3];
-
-    for i in 0..3 {
-        let bary = compute_spherical_barycentric(
-            sub_vertices_3d[i],
-            face_vertices_3d[0],
-            face_vertices_3d[1],
-            face_vertices_3d[2],
-        );
-
-        result[i] = (
-            face_2d[0].0 * bary.0 + face_2d[1].0 * bary.1 + face_2d[2].0 * bary.2,
-            face_2d[0].1 * bary.0 + face_2d[1].1 * bary.1 + face_2d[2].1 * bary.2,
-        );
-    }
-
-    (result[0], result[1], result[2])
-}
-
-/// Affine transformation from one triangle to another
-pub fn affine_transform_triangle(
-    point: (f64, f64),
-    source_tri: [(f64, f64); 3],
-    dest_tri: [(f64, f64); 3],
-) -> (f64, f64) {
-    // Source vectors relative to source_tri[0]
-    let (ax, ay) = (source_tri[1].0 - source_tri[0].0, source_tri[1].1 - source_tri[0].1);
-    let (bx, by) = (source_tri[2].0 - source_tri[0].0, source_tri[2].1 - source_tri[0].1);
-
-    // Destination vectors relative to dest_tri[0]
-    let (cx, cy) = (dest_tri[1].0 - dest_tri[0].0, dest_tri[1].1 - dest_tri[0].1);
-    let (dx, dy) = (dest_tri[2].0 - dest_tri[0].0, dest_tri[2].1 - dest_tri[0].1);
-
-    // Point relative to source_tri[0]
-    let (px, py) = (point.0 - source_tri[0].0, point.1 - source_tri[0].1);
-
-    // Solve: [ax bx] [s]   [px]
-    //        [ay by] [t] = [py]
-    let det = ax * by - bx * ay;
-    let s = (px * by - bx * py) / det;
-    let t = (ax * py - px * ay) / det;
-
-    // Apply same s,t to destination
-    let x = dest_tri[0].0 + s * cx + t * dx;
-    let y = dest_tri[0].1 + s * cy + t * dy;
-
-    (x, y)
-}
-// @TODO - needs to be added to spherical geometry, the other function there is not behaving correctly
-pub fn compute_spherical_barycentric(
-    point: Vector3D,
-    v0: Vector3D,
-    v1: Vector3D,
-    v2: Vector3D,
-) -> (f64, f64, f64) {
-    let total_area = spherical_triangle_area([v0, v1, v2]).unwrap();
-    let area0 = spherical_triangle_area([point, v1, v2]).unwrap();
-    let area1 = spherical_triangle_area([v0, point, v2]).unwrap();
-    let area2 = spherical_triangle_area([v0, v1, point]).unwrap();
-
-    (area0 / total_area, area1 / total_area, area2 / total_area)
-}
-
-/// Convert 2D Cartesian coordinates to barycentric coordinates
-pub fn cartesian_2d_to_barycentric(
-    point: (f64, f64),
-    triangle: [(f64, f64); 3],
-) -> (f64, f64, f64) {
-    let [f0, f1, f2] = triangle;
-    let (p_x, p_y) = point;
-
-    // // Check if point is very close to any vertex (handle special case)
-    // let vertex_tolerance = 1e-12;
-    // if (p_x - v0.0).length() < vertex_tolerance {
-    //     return (1.0, 0.0, 0.0);
-    // }
-    // if (point - v1).length() < vertex_tolerance {
-    //     return Some((0.0, 1.0, 0.0));
-    // }
-    // if (point - v2).length() < vertex_tolerance {
-    //     return Some((0.0, 0.0, 1.0));
-    // }
-
-    // Vectors from f0 to other vertices
-    let v0 = (f1.0 - f0.0, f1.1 - f0.1);
-    let v1 = (f2.0 - f0.0, f2.1 - f0.1);
-    let v2 = (p_x - f0.0, p_y - f0.1);
-
-    // Dot products
-    let d00 = v0.0 * v0.0 + v0.1 * v0.1;
-    let d01 = v0.0 * v1.0 + v0.1 * v1.1;
-    let d11 = v1.0 * v1.0 + v1.1 * v1.1;
-    let d20 = v2.0 * v0.0 + v2.1 * v0.1;
-    let d21 = v2.0 * v1.0 + v2.1 * v1.1;
-
-    let denom = d00 * d11 - d01 * d01;
-
-    if denom.abs() < 1e-10 {
-        // Degenerate triangle - return invalid
-        return (f64::NAN, f64::NAN, f64::NAN);
-    }
-
-    let bary_v = (d11 * d20 - d01 * d21) / denom; // Weight for f1
-    let bary_w = (d00 * d21 - d01 * d20) / denom; // Weight for f2
-    let bary_u = 1.0 - bary_v - bary_w; // Weight for f0
-
-    (bary_u, bary_v, bary_w)
 }
